% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{nbr_pres_knn}
\alias{nbr_pres_knn}
\title{Neighborhood Preservation Between Nearest Neighbor Matrices}
\usage{
nbr_pres_knn(kin, kout, k = ncol(kin), n_threads = 0)
}
\arguments{
\item{kin}{Nearest neighbor matrix. The "ground truth" or reference indices.}

\item{kout}{Nearest neighbor matrix. A set of distances to compare to the reference
indices.}

\item{k}{The size of the neighborhood, where k is the number of neighbors to
include in the neighborhood.}

\item{n_threads}{the maximum number of threads to use. \code{0} or \code{1} runs
serially.}
}
\value{
Vector of preservation values, one for each row of \code{kin}.
}
\description{
Calculates the neighborhood preservation for each observation in a dataset,
represented by two matrices of the indices of the nearest neighbors. The
first matrix is the "ground truth", the second being the estimation or
approximation. The neighborhood preservation is calculated for each row where
each element \code{d[i, k]} is taken to be the index of the kth nearest neighbor
of \code{i}.
}
\details{
Approximate nearest neighbor methods, e.g.
\href{https://cran.r-project.org/package=RcppAnnoy}{RcppAnnoy}, can find
k-nearest neighbors quite efficiently and so makes calculating preservation
values for larger datasets feasible.

The neighborhood preservation can vary between 0 (no neighbors in common)
and 1 (perfect preservation). For nearest-neighbor matrices that exclude
self-neighbors, random performance gives an approximate value of k / (n - 1),
where k is the size of the neighborhood and n is the number of observations
or items in the dataset.
}
