% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadra-package.R
\docType{package}
\name{quadra-package}
\alias{quadra}
\alias{quadra-package}
\title{quadra: QUality Assessment of Dimensonality Reduction Algorithms}
\description{
Package for quantifying the embeddings produced by dimensionality reduction methods.
}
\author{
\strong{Maintainer}: James Melville \email{jlmelville@gmail.com}

}
\keyword{internal}
